refactor(user): Change idUser type to Long
- Update idUser to Long in entities and DTOs
- Adapt services and controllers to use Long for user IDs

+++ src/main/java/com/github/martinalexis/course_management/user/controller/v1/UserController.java
@@ -57,7 +57,7 @@ public class UserController {
                             examples = @ExampleObject(name = "Internal Error", value = GlobalExceptionJsonExamples.UNEXPECTED_ERROR_RESPONSE))
             )
     })
-    public ResponseEntity<UserResponseDto> getUserById(@PathVariable("id") Integer id) {
+    public ResponseEntity<UserResponseDto> getUserById(@PathVariable("id") Long id) {
         UserResponseDto user = userUseCase.getById(id);
         return ResponseEntity.status(HttpStatus.OK).body(user);
     }
